Added keywords + refactored constants files and references and unit tests
