+++ src/app/constants/keywords.constants.ts
@@ -1,0 +1,1 @@
+export const KEYWORDS = new Set(['AND', 'OR', 'NOT', 'TRUE', 'FALSE']);
